+++ packages/react-article-components/src/constants/anchor.js
@@ -1,0 +1,1 @@
+export const relatedPostAnchor = 'related-post-anchor'