+++ packages/index-page/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.16-rc.4](https://github.com/twreporter/twreporter-npm-packages/compare/@twreporter/index-page@1.2.16-rc.3...@twreporter/index-page@1.2.16-rc.4) (2022-07-04)
+
+**Note:** Version bump only for package @twreporter/index-page
+
+
+
+
+
+## [1.2.16-rc.3](https://github.com/twreporter/twreporter-npm-packages/compare/@twreporter/index-page@1.2.16-rc.2...@twreporter/index-page@1.2.16-rc.3) (2022-07-01)
+
+**Note:** Version bump only for package @twreporter/index-page
+
+
+
+
+
+## [1.2.16-rc.2](https://github.com/twreporter/twreporter-npm-packages/compare/@twreporter/index-page@1.2.16-rc.1...@twreporter/index-page@1.2.16-rc.2) (2022-06-30)
+
+**Note:** Version bump only for package @twreporter/index-page
+
+
+
+
+
+## [1.2.16-rc.1](https://github.com/twreporter/twreporter-npm-packages/compare/@twreporter/index-page@1.2.16-rc.0...@twreporter/index-page@1.2.16-rc.1) (2022-06-14)
+
+**Note:** Version bump only for package @twreporter/index-page
+
+
+
+
+
+## [1.2.16-rc.0](https://github.com/twreporter/twreporter-npm-packages/compare/@twreporter/index-page@1.2.15...@twreporter/index-page@1.2.16-rc.0) (2022-06-08)
+
+**Note:** Version bump only for package @twreporter/index-page
+
+
+
+
+
 ## [1.2.15](https://github.com/twreporter/twreporter-npm-packages/compare/@twreporter/index-page@1.2.15-rc.6...@twreporter/index-page@1.2.15) (2022-06-07)
 
 **Note:** Version bump only for package @twreporter/index-page